MA-19449 Add VPU driver in TEE
1. complete the smc handler for VPU encoder and decoder
2. When the VPU starts decoding, it will first check whether the values
of all registers that need to be protected are all secure memory or normal
memory. If the align condition is met, the values of the protected
registers in the cache are written into the registers. If it is secure
memory, then set the VPU to domain2, otherwise it is domain0.

Change-Id: I367c366a67055c537dc270e0c1a5201f3a81e9ae
Signed-off-by: zhai.he <zhai.he@nxp.com>