MA-20178 Add DCSS driver into TEE
Because secure UI is enabled, write operations on DCSS
registers on the linux side need to be protected.
Therefore, write operations of DCSS are excuted in TEE,
which is completed by fastcall between the linux side and TEE.

And when start confirmation ui, it should be init submodule's
regs of dcss for showing TUI, and regs of primary plane will be
protected when TUI displaying.

Change-Id: I5a13c225109ae42adf817e2c5b41ce3fbad988cb
Signed-off-by: zhai.he <zhai.he@nxp.com>