MA-20178-2 Modify the restrictions of the address register of ctxld
Becasue when confirmation TA is loaded, ctxld_db and ctxld_sb buffer
will use TEE memory, when display updated, ctxld_db and ctxld_sb will be
writted into 0x32e23010 and 0x32e23018 register.

Only when the physical addresses of ctxld_db and ctxld_sb are written to
these two registers, the register can be allowed to be written,
otherwise it is not allowed to write.

Change-Id: I39d095469f16b7ced0db8245e9c7f2e9f1271f6b
Signed-off-by: zhai.he <zhai.he@nxp.com>