MA-20178-3 Add primary plane of dtg protection and change wait ctxld condition
There are primary plane related controls in the DTG component, so these
registers need to be protected.

And since when ctxld is idle, ctxld_enable will be clear, so can use
this flag to determine whether ctxld can be used.

when initial secure UI, no need to negate the index of ctxld_buffer, because the
Linux side will negate index after processing ctxld and pass it to TEE.

Change-Id: I5df988c3913d5884fb432c399c476db4a02b00ee
Signed-off-by: zhai.he <zhai.he@nxp.com>