MA-21257-3 add imx93 target and ele driver
add imx93 target and the ele/mu drivers. Due to the MU
is reused by both secure and non-secure worlds, race may
happen if we call the ELE APIs at linux runtime.

To avoid the MU conflict between secure and non-secure, this
commit will not always derive the RNG from ELE. Instead, we
only seed the rand() at the begining. It is not the best
solution but we have no better choice before we figure out
how to avoid the conflict.

Change-Id: I97d9ad013d3ef5b46c6f022eae0dd0f130138f5d
Signed-off-by: Ji Luo <ji.luo@nxp.com>