MA-21549 Fix the issue that Android GTS 1080p-60fps performance failed
Because the trusty kernel uses memcpy in libc by default, the memcpy does
not do too much optimization, so the optimized memcpy provided by ARM is
used, and an alignment fault occurs because the secure heap mmap becomes
uncached, so in order to avoid this problem, Use the cached attribute to
map.

Note: Trusty only supports three attributes by default: cached, uncached
(nGnRnE), uncached device (nGnRE), and the default is nGRE when mapping
uncached memory on the linux side, alignment will not occur when dealing
with non-aligned memcpy, so memory needs to be Mapped to cached to avoid
alignment fault.

The memcpy_aarch64.S is obtained from the following url:
https://github.com/ARM-software/optimized-routines/blob/master/string/aarch64/memcpy.S

Change-Id: I9d7f5a82b09f6e1207fb913f82c73f53dcaa5f7a
Signed-off-by: He Zhai <zhai.he@nxp.com>