MA-21705-2 use software backed rng before hardware ready
At the very early boot stage, thread_init_early() would call
thread_init_cookie() to generate and bind the thread to a cookie
which is a 64 bits random. The 64 bits random was generated
from rand_get_bytes() which would call platfrom provided RNG
which is CAAM/ELE on i.MX.

However, at the very early stage, the heap is not ready and
RNG hardware on i.MX can't be used. Instead, this commit uses
the software backed RNG and seed the RNG with current system
timer.

Change-Id: Ic4394236da5820901b453bf2e5c55a0eba356fa4
Signed-off-by: Ji Luo <ji.luo@nxp.com>